+++ packages/mds-jurisdiction-service/package.json
@@ -22,7 +22,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@mds-core/mds-logger": "0.1.24",
-    "@mds-core/mds-orm": "0.1.0",
+    "@mds-core/mds-repository": "0.1.0",
     "@mds-core/mds-schema-validators": "0.1.2",
     "@mds-core/mds-service-helpers": "0.1.0",
     "@mds-core/mds-types": "0.1.23",
