+++ packages/mds-repository/naming-strategies/index.ts
