+++ packages/mds-repository/naming-strategies/mds-naming-strategy.ts
