+++ packages/mds-repository/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mds-core/mds-orm",
+  "name": "@mds-core/mds-repository",
   "version": "0.1.0",
   "description": "Mobility Data Specification ORM",
   "main": "dist/index.js",
