+++ packages/mds-repository/tests/mds-naming-strategy.spec.ts
